requirements.txt
python-telegram-bot[job-queue]==21.6
python-dotenv==1.0.1
pypdf==5.1.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
